[skip ci] 🛡️ Sentinel: [CRITICAL] Fix critical vulnerabilities and IDORs
This commit addresses several critical security vulnerabilities found in the codebase:
- Removed hardcoded fallback for JWT_SECRET and enforced strict presence checks to prevent auth bypass.
- Added replay attack prevention to the Sign In With Solana (SIWS) authentication route by tracking and rejecting used nonces.
- Enforced strict signature validation on Lemon Squeezy webhooks and removed the hardcoded dev secret.
- Fixed IDOR on `/api/runs` by explicitly filtering runs by the authenticated user's ID.
- Fixed IDOR on `/api/vault/[address]/activity` by verifying that the authenticated user is a member of the team that owns the requested vault.
- Fixed CI build failures related to Next.js route handler typings by refactoring non-route exports to a separate module.
